Added ToggleArmCoast
In case we need to get new positions for the motors (e.g. new CubePlaceMid positions)
